requirements.txt
altair==5.1.1
annotated-types==0.5.0
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
decorator==5.1.1
elevenlabs==0.2.26
exceptiongroup==1.1.3
executing==1.2.0
gitdb==4.0.10
GitPython==3.1.36
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
numpy==1.25.2
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rfc3986==1.5.0
rich==13.5.2
rpds-py==0.10.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
streamlit==1.26.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.10.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.22.0
wcwidth==0.2.6
websockets==11.0.3
zipp==3.16.2
